pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Alex Crouzen <alexander@crouzen.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
"discord.py" = "^1.7.3"
APScheduler = "^3.9.1"
parse = "^1.19.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"